add full unit tests for service classes; a bunch of bug fix and code optimization

+++ xuecheng-edu-content/xuecheng-edu-content-model/src/main/java/com/xuecheng/content/model/po/CourseMarket.java
@@ -27,12 +27,12 @@ public class CourseMarket implements Serializable {
     /**
      * 现价
      */
-    private Float price;
+    private Double price;
 
     /**
      * 原价
      */
-    private Float originalPrice;
+    private Double originalPrice;
 
     /**
      * 咨询qq
